global: restructure the layout for simplicity
The individual python files themselves are pretty small, so I don't see
the need to split them all into their own sub-module.

Secondly, for easy of integration, we can just make one CFFI wrapper
binding library that links to all the tss2 libraries. Theirs no need to
complicate this with multiple CFFI builds and modules. We just make this
one monolithic beast, which greatly simplifies the build and deployment
model.
