+++ docs/esys.rst
@@ -1,6 +1,6 @@
 Esys
 ====
 
-.. autoclass:: tpm2_pytss.pyesys.EsysContext
+.. autoclass:: tpm2_pytss.ESAPI
    :members:
    :undoc-members:
