--- tpm2_pytss/pyesys/__init__.py
@@ -1,7 +1,0 @@
-"""
-SPDX-License-Identifier: BSD-3
-"""
-
-from .types import *
-from .EsysContext import EsysContext
-from .TctiLdr import TctiLdr
